app/requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
tomli==1.1.0
tzlocal==2.1
